Fix hangs by using the `try_io` API
The new test case included in this change would reliably hang; at
least on Linux with `epoll`. I have the hunch that this may be related
to spurious wakeups and/or everything being edge-triggered now.

Adapting the code to use looping, and the `try_io` helper method, as
illustrated in the current documentation for `AsyncFd`, seems to fix
the hangs.